refactor: BettingMoney 패키지 선언을 betting으로 수정
