+++ src/main/java/blackjack/domain/betting/BettingMoney.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.betting;
 
 public class BettingMoney {
